backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
jaclang>=0.9.3
byllm>=0.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
httpx==0.25.1
pytest==7.4.3
google-generativeai
jaseci==1.4.0
jaseci-ai-kit==1.4.0
jaseci-serv==1.4.0